requirements_test.txt
coverage==7.4.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
freezegun==1.4.0
factory-boy==3.3.0
pytest-mock==3.12.0
# snapshottest==0.5.1
pytest-sugar==0.9.7
